refactor

+++ operators/multiclusterobservability/controllers/placementrule/placementrule_controller.go
@@ -49,13 +49,9 @@ import (
 )
 
 const (
-	ownerLabelKey                                  = "owner"
-	ownerLabelValue                                = "multicluster-observability-operator"
 	managedClusterObsCertName                      = "observability-managed-cluster-certs"
 	nonOCP                                         = "N/A"
 	disableAddonAutomaticInstallationAnnotationKey = "addon.open-cluster-management.io/disable-automatic-installation"
-	MulticlusterGlobalHubAgentName                 = "multicluster-global-hub-agent"
-	MulticlusterGlobalHubAgentNamespace            = "multicluster-global-hub-agent-namespace"
 )
 
 var (
@@ -161,6 +157,11 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 		return ctrl.Result{}, err
 	}
 
+	reqLogger.Info("Coleen To list observabilityaddon", "namespace", req.Namespace)
+	for _, ep := range obsAddonList.Items {
+		reqLogger.Info("Coleen To list observabilityaddon", "namespace", ep.Namespace)
+	}
+
 	if !deleteAll {
 		if err := createAllRelatedRes(
 			r.Client,
@@ -203,14 +204,14 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 	}
 	for _, work := range workList.Items {
 		if work.Name != work.Namespace+workNameSuffix {
-			reqLogger.Info("To delete invalid manifestwork", "name", work.Name, "namespace", work.Namespace)
+			reqLogger.Info("Coleen To delete invalid manifestwork", "name", work.Name, "namespace", work.Namespace)
 			err = deleteManifestWork(r.Client, work.Name, work.Namespace)
 			if err != nil {
 				return ctrl.Result{}, err
 			}
 		}
 		if !slices.Contains(latestClusters, work.Namespace) {
-			reqLogger.Info("To delete manifestwork", "namespace", work.Namespace)
+			reqLogger.Info("Coleen To delete manifestwork", "namespace", work.Namespace)
 			err = deleteManagedClusterRes(r.Client, work.Namespace)
 			if err != nil {
 				return ctrl.Result{}, err
@@ -232,6 +233,65 @@ func (r *PlacementRuleReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 			}
 		} else {
 			staleAddons = commonutil.Remove(staleAddons, mcaddon.Namespace)
+		}
+	}
+
+	clusterType := ""
+	isSNO, err := isSNO(ctx, r.Client)
+	if err == nil && isSNO {
+		clusterType = "SNO"
+	}
+	clusterID := "hub-cluster"
+	_, err = getClusterID(ctx, r.Client)
+	if err != nil {
+		// to differentiate ocp 3.x
+		clusterType = "ocp3"
+	}
+	params := HubCollectorParams{
+		isUWL:        false,
+		clusterID:    clusterID,
+		clusterType:  clusterType,
+		obsAddonSpec: *mco.Spec.ObservabilityAddonSpec,
+		//hubInfo:      hubInfo,
+		//allowlist:    list,
+		replicaCount: 1,
+		nodeSelector: mco.Spec.NodeSelector,
+		tolerations:  mco.Spec.Tolerations,
+	}
+	metricsCollectorDeployment, err := GenerateMetricsCollectorForHub(ctx, mco, params)
+	if err != nil {
+		log.Error(err, "Failed to generate metrics collector deployment")
+		return ctrl.Result{}, err
+	}
+
+	found := &appsv1.Deployment{}
+	err = r.Client.Get(ctx, types.NamespacedName{Name: metricsCollectorName,
+		Namespace: config.GetDefaultNamespace()}, found)
+	if err != nil {
+		if k8serrors.IsNotFound(err) {
+			err = r.Client.Create(ctx, metricsCollectorDeployment)
+			if err != nil {
+				log.Error(err, "Failed to create deployment", "name", metricsCollectorName)
+				return ctrl.Result{}, err
+			}
+			log.Info("Created deployment ", "name", metricsCollectorName)
+		} else {
+			log.Error(err, "Failed to check the deployment", "name", metricsCollectorName)
+			return ctrl.Result{}, err
+		}
+	} else {
+		if !reflect.DeepEqual(metricsCollectorDeployment.Spec.Template.Spec, found.Spec.Template.Spec) ||
+			!reflect.DeepEqual(metricsCollectorDeployment.Spec.Replicas, found.Spec.Replicas) {
+			metricsCollectorDeployment.ObjectMeta.ResourceVersion = found.ObjectMeta.ResourceVersion
+			if found.Status.ReadyReplicas != 0 {
+				metricsCollectorDeployment.Spec.Template.ObjectMeta.Labels[restartLabel] = time.Now().Format("2006-1-2.1504")
+			}
+			err = r.Client.Update(ctx, metricsCollectorDeployment)
+			if err != nil {
+				log.Error(err, "Failed to update deployment", "name", metricsCollectorName)
+				return ctrl.Result{}, err
+			}
+			log.Info("Updated deployment ", "name", metricsCollectorName)
 		}
 	}
 
@@ -345,7 +405,7 @@ func createAllRelatedRes(
 		currentClusters = commonutil.Remove(currentClusters, managedCluster)
 		if isReconcileRequired(request, managedCluster) {
 			log.Info(
-				"Monitoring operator should be installed in cluster",
+				"Coleen Monitoring operator should be installed in cluster",
 				"cluster_name",
 				managedCluster,
 				"request.name",
@@ -377,22 +437,15 @@ func createAllRelatedRes(
 	}
 	managedClusterListMutex.RUnlock()
 
-	// create managed cluster resource in the namespace of mco
-	log.Info("To create managedcluster resources", "namespace", request.Namespace)
-	if err = createManagedClusterRes(c, mco,
-		"local-cluster", "local-cluster",
-		works, metricsAllowlistConfigMap, crdv1Work, endpointMetricsOperatorDeploy, hubInfoSecret, false); err != nil {
-		failedCreateManagedClusterRes = true
-		log.Error(err, "Failed to create managedcluster resources", "namespace", request.Namespace)
-	}
-
 	failedDeleteOba := false
 	for _, cluster := range currentClusters {
 		log.Info("Coleen To delete observabilityAddon", "namespace", cluster)
-		err = deleteObsAddon(c, cluster)
-		if err != nil {
-			failedDeleteOba = true
-			log.Error(err, "Failed to delete observabilityaddon", "namespace", cluster)
+		if cluster != config.GetDefaultNamespace() {
+			err = deleteObsAddon(c, cluster)
+			if err != nil {
+				failedDeleteOba = true
+				log.Error(err, "Failed to delete observabilityaddon", "namespace", cluster)
+			}
 		}
 	}
 
@@ -553,8 +606,6 @@ func (r *PlacementRuleReconciler) SetupWithManager(mgr ctrl.Manager) error {
 
 	// Watch changes for AddonDeploymentConfig
 	AddonDeploymentPred := GetAddOnDeploymentPredicates()
-
-	mcghAgentDeploymentPred := GetMCGHAgentDeploymentPredicates()
 
 	obsAddonPred := predicate.Funcs{
 		CreateFunc: func(e event.CreateEvent) bool {
@@ -844,10 +895,7 @@ func (r *PlacementRuleReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		Watches(&source.Kind{Type: &corev1.Secret{}}, &handler.EnqueueRequestForObject{}, builder.WithPredicates(certSecretPred)).
 
 		// secondary watch for alertmanager accessor serviceaccount
-		Watches(&source.Kind{Type: &corev1.ServiceAccount{}}, &handler.EnqueueRequestForObject{}, builder.WithPredicates(amAccessorSAPred)).
-
-		// secondary watch for MultiClusterGlobalHub deployment
-		Watches(&source.Kind{Type: &appsv1.Deployment{}}, &handler.EnqueueRequestForObject{}, builder.WithPredicates(mcghAgentDeploymentPred))
+		Watches(&source.Kind{Type: &corev1.ServiceAccount{}}, &handler.EnqueueRequestForObject{}, builder.WithPredicates(amAccessorSAPred))
 
 	// watch for AddonDeploymentConfig
 	if _, err := r.RESTMapper.RESTMapping(schema.GroupKind{Group: addonv1alpha1.GroupVersion.Group, Kind: "AddOnDeploymentConfig"}, addonv1alpha1.GroupVersion.Version); err == nil {
